requirements.txt
fastapi
uvicorn
pywin32
psycopg[binary]
python-dotenv
